Kernel: Split the FATFileSystem.{cpp,h} files into smaller components
